fix(OneDataCollection): use source defaults when resetting Sort by and Group by
