+++ packages/react/src/patterns/OneDataCollection/__tests__/index.spec.tsx
@@ -1283,118 +1283,4 @@ describe("Collections", () => {
       expect(screen.getByText("user20@example.com")).toBeInTheDocument()
     })
   })
-
-  test("Reset to default restores defaultSortings and defaultGrouping", async () => {
-    type Person = {
-      id: number
-      name: string
-      department: string
-    }
-
-    const mockData: Person[] = [
-      { id: 1, name: "John Doe", department: "Engineering" },
-      { id: 2, name: "Alice Brown", department: "Product" },
-    ]
-
-    const { result } = renderHook(
-      () =>
-        useDataCollectionSource<
-          Person,
-          FiltersDefinition,
-          SortingsDefinition,
-          SummariesDefinition,
-          ItemActionsDefinition<Person>,
-          NavigationFiltersDefinition,
-          GroupingDefinition<Person>
-        >({
-          dataAdapter: {
-            fetchData: async () => ({ records: mockData }),
-          },
-          sortings: {
-            name: { label: "Name" },
-            department: { label: "Department" },
-          },
-          defaultSortings: { field: "name", order: "desc" },
-          grouping: {
-            groupBy: {
-              department: {
-                name: "Department",
-                label: (groupId) => String(groupId),
-              },
-            },
-          },
-        }),
-      { wrapper: TestWrapper }
-    )
-
-    render(
-      <TestWrapper>
-        <OneDataCollection
-          source={result.current}
-          visualizations={[
-            {
-              type: "table",
-              options: {
-                columns: [
-                  {
-                    label: "Name",
-                    render: (item: Person) => item.name,
-                    sorting: "name" as const,
-                  },
-                  {
-                    label: "Department",
-                    render: (item: Person) => item.department,
-                  },
-                ],
-              },
-            },
-          ]}
-        />
-      </TestWrapper>
-    )
-
-    expect(result.current.currentSortings).toEqual({
-      field: "name",
-      order: "desc",
-    })
-    expect(result.current.currentGrouping).toBeUndefined()
-
-    await act(async () => {
-      result.current.setCurrentSortings({ field: "name", order: "asc" })
-      result.current.setCurrentGrouping({ field: "department", order: "asc" })
-    })
-
-    await waitFor(() => {
-      expect(result.current.currentSortings).toEqual({
-        field: "name",
-        order: "asc",
-      })
-      expect(result.current.currentGrouping).toEqual({
-        field: "department",
-        order: "asc",
-      })
-    })
-
-    const settingsButton = await screen.findByRole("button", {
-      name: /settings/i,
-    })
-    await act(async () => {
-      await userEvent.click(settingsButton)
-    })
-
-    const resetButton = await screen.findByRole("button", {
-      name: /reset to default/i,
-    })
-    await act(async () => {
-      await userEvent.click(resetButton)
-    })
-
-    await waitFor(() => {
-      expect(result.current.currentSortings).toEqual({
-        field: "name",
-        order: "desc",
-      })
-      expect(result.current.currentGrouping).toBeUndefined()
-    })
-  })
 })
